Change Image checkbox to input checkbox
